requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
Flask==2.0.1
Flask-Login==0.5.0
gunicorn==20.1.0
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
mariadb==1.0.7
MarkupSafe==2.0.1
python-dotenv==0.18.0
requests==2.25.1
urllib3==1.26.6
Werkzeug==2.0.1
